refactor validate to correct helper location
